build: LP64 (gcc) で -Wformat を有効にする (実在した 9 箇所を直した)
**`%d` に `long` を渡すのは LP64 では本物の間違い**だが、gcc の枝では
`-Wno-format` で塞いであった。実在した 9 箇所を直してから有効にした。

**LP64 では `%d` は下位 32bit しか読まない。** たとえば 2GB を超えるファイルの
読み込み進捗が「Reading -1234/…」のような数になる (`src/core/fileio.cc`)。他は
行番号・桁・バッファ版数・マーカ位置・前置引数で、実用上の値では表に出ない。
`%ld` は Windows (LLP64、`long` は 32bit) でも正しいので、枝を分けずに直せる。

**`-Wno-format` を外すだけでは効かない。** gcc は `-Wformat` を既定で有効にして
いない (`-Wall` に入っている) ので、外しても何も出ない。**一度それで「0 件
だった」と誤解した** — 測ったつもりで測れていなかった。clang は既定で有効なので、
あちらは `-Wno-format` が効いている。

Windows 側 (MSVC の `/wd4477`、Clang の `-Wno-format`) は**塞いだまま**にする:
LLP64 では `%d` + `long` が**すべて偽陽性**になる (実測で 100 件以上)。**同じ
警告が片方では役に立ち、もう片方では雑音になる。**

`-Wformat-truncation` は塞いだ。出るのは `snprintf` を境界付きで正しく使っている
6 箇所で、「入り切らないかもしれない」という指摘である (溢れはしない)。

**`-Wall` を丸ごと入れるのはやめた。** 測ったら 396 件出て、大半は
`-Wnarrowing` 60 / `-Wsign-compare` 51 / `-Wparentheses` 42 のようなこの
コードベースの書き方そのものだった。残り (`-Wstrict-aliasing` 12、`-Wswitch` 6、
`-Wclass-memaccess` 1 ほか) は issue に分けた。

実測: `_build/linux` を**コンテナの中から**消して作り直し、警告はリンカの
`-z stacksize` の注意 1 件だけ。linux 1184 件 / 既知失敗 17、llvm-mingw x86_64
1198 件 / 7。どちらも `no failures outside misc/known-failures`。

+++ src/core/lprint.cc
@@ -1539,7 +1539,7 @@ print_buffer_name (wStream &stream, const Buffer *bp)
   if (bp->b_version != 1)
     {
       char b[64];
-      sprintf (b, "<%d>", bp->b_version);
+      sprintf (b, "<%ld>", bp->b_version);
       stream.add (b);
     }
 }
@@ -1577,7 +1577,7 @@ print_marker (wStream &stream, const print_control &, lisp object)
       else
         {
           char b[64];
-          sprintf (b, ": %u", xmarker_point (object));
+          sprintf (b, ": %ld", (long)xmarker_point (object));
           stream.add (b);
         }
       stream.add ('>');
