+++ src/gen-syms.cc
@@ -3227,7 +3227,7 @@ process_interactive ()
 static void
 print_version ()
 {
-  printf ("int dump_version = %d;\n", time (0));
+  printf ("int dump_version = %d;\n", (int)time (0));
 }
 
 void
